On unstable branch, disabled do_mmap_ckpt_image; will look for bug fix for that later.
Also cleaned up code in unstable to remove some error messages, add "static"
to some functions so as to not pollute user space too much.

Also fixed some things to try to get ./configure --disaple-pid-virtualization
working again in unstable.  It now compiles and checkpoints, but fails on
restart.  The error message happens when a call to clone() fails:
  Line 4058 in mtcp/mtcp.c is reporting error 22 (invalid arg) from
  clone_entry.
Will analyze this problem later.