MTCP now checkpoints signal handlers if not running under DMTCP.
Standalone MTCP has not been checkpointing signal handlers ever since DMTCP was
created.  However, this functionality has now been restored.  Currently, if
DMTCP is present, DMTCP will handle signals, but it may make more sense for
MTCP to always provide this functionality, as signals are local to a single
process.

A new variable has also been added to mtcp.c called dmtcp_exists.  This variable
is set by DMTCP, and tells MTCP that it is running under DMTCP.  This is a much
more scalable approach to this problem, as it allows both DMTCP and standalone
MTCP to exist simultaneously on a single system (as opposed to the old approach
of using a preprocessor conditional).