Re-evaluate coordinator state on client disconnect
After the coordinator has sent out SUSPEND message, if a client dies
(normal exit or exit due to an error), the coordinator should
re-evaluate the minimumState() and proceed to set the minimum state to
SUSPENDED if all other processes are already suspended.

Earlier, the coordinator would stuck at a point where all the processes
are SUSPENDED, but the coordinator didn't proceed towards the next
barrier.