Fixes for mtcp standalone
Rev 1296 introduced some new callback between mtcp and dmtcp. However,
when calling the callbacks, the code was not checking whether the
callbacks are properly initialized. In case of standalone MTCP, the
callbacks are initialized to NULL, and thus it produced a SEGFAULT. This
commit provides the proper check for callbacks.