Fixes for pthread_join() wrapper family
pthread_join() is a blocking call which waits for the given thread to
exit. However, if the process is checkpointed in the middle of
pthread_join(), on restart, pthread_join() would never return and would
spin on a futex syscall indefinitely causing it to use 100% cpu. The
reason for the spin is the changed tid of the thread after restart.

This checkin fixes this issue by replacing the call to
_real_pthread_join() with _real_pthread_tryjoin_np() so that the thread
is never blocked. In order to satisfy the pthread_join() semantics, some
tweaks were needed. Refer to code in pthread_join() wrapper for more
details.