restore_term_settings: add warning message about SIGWINCH and Apache
This patch adds a warning in restore_term_settings() for
users testing DMTCP with Apache. Users should disable
the SIGWINCH that DMTCP sends on restart because Apache
uses that as a signal for terminating gracefully.

Future patches can include a method to do this automatically
by wrapping the call to send SIGWINCH in an ifdef that gets
disabled with a configure option.