dmtcp-aarch64: use NEXT_FNC_DEFAULT for pthread functions
The pthread functions in syscallsreal.c use
dlvsym(RTLD_NEXT, v2.4, ...) to get to the _real_ functions in
libpthread, but the version numbers on the latest
distributions have changed.  This patch changes the
functions to use NEXT_FNC_DEFAULT at run-time.

This is a temporary fix. A cleaner solution would be to fix
the initialization code in initialize_libpthread_wrappers().