Fixed #56: Segfault when restoring heap on restart.
During restart, we try to restore the heap as it existed at the time of
ckpt.  A JNOTE was trying to print the original and current values of
brk. The variable holding the current brk (curBrk) was of type 'char*'
and so JNOTE tried to print the corresponding string on the console.
Since the current heap had been unmapped by mtcp_restart, accessing
"curBrk" generated a segfault. The fix is to change the type of
curBrk to "void*", thus forcing JNOTE to print the value of curBrk
instead of dereferencing curBrk itself.