fileconnection: add check to verify the existence of a dir
This patch is a workaround for what is arguably a bug in the Luster
FS. See: https://jira.hpdd.intel.com/browse/LU-6528.  When multiple
independent processes call mkdir() on an existing directory,
a process can sometimes get an EACCES error, which causes our
createDirectoryStructure function to assert. This patch adds an
extra check (using the stat() call) in createDirectoryStructure in
case the call to mkdir() returned EACCES.