PID: Removed definitions of gettid() and friends.
These symbols are not exported by libc and so we don't need to export
them either. This also caused segfault if the application binary that
was compiled with '-rdynamic' provided a definition of 'gettid()' (as
reported in #129).

Also added gettid/getpid test to syscall-tester and added -rdynamic
compilation flag along with creating a definition for gettid().

Fixes #129.