Removed no-dependencies flag from configure.ac.
This allows automake to do dependency tracking (i.e., if a header file
is changed, all files #including it will be recompiled). For this, the
'subdir-objects' flags had to be disabled due to a bug in automake 1.14.
The fix for this bug would be released as part of automake 1.16. At that
point, we can add 'subdir-objects' back.