Added a new interface to the plugin API.
This introduces a new struct "DmtcpPluginDescriptor_t", that would be
passed on to the plugin engine during plugin initialization. This data
structure allows the plugins to register their event hook with the
plugin engine along with some other metadata (name, author, etc.) about
the plugin.

The plugin engine would use the event hooks registered by various
plugins to provide the event notification in the required order. The
order of event notification would be reversed for post-ckpt/restart
events.