FILE: Better handling of deleted files.
One some systems, deleted files have " (deleted)" as a prefix whereas on
other systems, this is a suffix to the filename (when viewed through
"ls -l /proc/PID/fd/" or "cat /proc/PID/maps").

This fix uses "fstat" system call to check the number of hard links to
the given file. An unlinked file has "0" hardlinks.