Fixes #45: Bug fix for sem_wait and glibc >= 2.21.
In glibc >= 2.21, sem_wait registers a cleanup handler before blocking
for the semaphore. The handler is popped before returning from it.
However, if we use it in stopthisthread(), the cleanup handler is not
popped up on restart because, we do a longjmp. Thus if a thread, that was
created before ckpt, were to do pthread_exit, the ptherad_exit routine
will pop all pending cleanup handlers and try to execute them. At this
point the stale sem_wait() handler results in SEGFAULT (since the
semaphore is invalid now).

This was most noticable on pthread2/3 tests since they have threads that
are created before ckpt and removed after restart.