Exec: Always make a copy of the incoming envp.
Sometimes the caller just sends "environ" as an argument to execve, etc.
Later on, if we modify the environment vector by adding a new
environment variable, environ might be relocated to a new address, thus
making the value passed into "execve" invalid. The fix is to always make
a local copy of the incoming environment vector and use it for all our
purposes.