Give frisbee some extra time to startup.
If the non-starter processes didn't have a chance to bind() yet, the
started would to connect and would terminate. The fix is to allow the
started process to retry connect (20 times with a 100 ms sleep
inbetween). This is a hackish fix, we need to provide a better guarantee
for such tests.