Fix updating root of process tree after parent exits
On older systems, a process is inherited by init (pid = 1) after its
parent dies. However, with the new per-user init process, the parent
pid is no longer "1"; it's the pid of the user-specific init process.

The current process is the root-of-process-tree, if the new
value of getppid() is different than the previous one.