PID: Fixed a race when parent exits right after fork.
The parent will set the virtual-pid environment variable with
"<virtual-PID>:<virtual-PPID>". The child then records a mapping between
_real_getpid() and virtual-PID as well as _real_getppid() and
virtual-PPID. However, if the parent dies before _real_getppid() is
called, the latter would return the real-pid of the new parent (init
process). The child process then records a mapping between virtual-PPID
of the original parent and real-pid of "init" process. Later on, the
getppid() wrapper fails to detect the exited parent and keeps returning
the virtual-pid of the original parent. This further prevents
ProcessInfo from marking the process as the root of process tree and
thus we see errors on restart.

The bug has been present for a while but didn't show up on faster
machines. We observed it on travis-ci's containerized setup where the
bug showed up quite frequently after a series of recent commits.

Closes #148.