MTCP: Fixed a race w/ unlinked shared-memory files.
For shared memory areas backed by unlinked files, MTCP used to create
the file with RW permissions and populate it with the contents from the
ckpt image. Next, it would close and reopen the file descriptor to
reflect the appropriate pre-ckpt flags. Finally, the memory area would
be mapped and the file would be unlinked.

When there are two or more processes involved, they all follow the same
steps. However, sometimes, one of the processes will unlink the file
right when a different process is in the middle of closing and reopening
the file descriptor. This would cause the reopening of the file to fail
and the process to exit.

The fix is to remove the closing/reopening logic since the flags on the
file descriptors don't matter. We are anyways creating the shared-memory
area with appropriate flags, so it's okay to have RW flags sets on the
file descriptor even if the shared memory area needs only the R flag.