mtcp_restart: fix the logic for mmap-ing truncated files
This patch fixes an issue reported by Ankit. The problem was that We
were hitting an mtcp_abort() in adjust_for_smaller_file_size() because
MTCP was failing to restore a memory region where the application had
mmap-ed /dev/zero, probably to emulate a fast calloc().

The original intent of the function was to figure out if the backing
file of an mmap-ed region has truncated since checkpoint time, compute
the difference, and then patch the memory region using MAP_ANONYMOUS.

The logic in the function is difficult to follow, but, more importantly,
it fails to correctly handle the situation when the size of the backing
file of an mmap-ed region is 0, which is the case with /dev/zero.

This patch simplifies the code and fixes this issue by: a) mmap-ing an
entire region as MAP_ANONYMOUS if its backing file has truncated since
checkpoint time, and b) avoiding calls to the function.