sysvipcwrappers: Fix {IPC|SEM}_INFO cases for semctl
An application can call semctl({IPC|SEM}_INFO) directly to
fetch the information about system-wide semaphore limits
and parameters. In this case, the values of the first two
arguments to semctl() -- semid and semnum -- are irrelevant,
and the application can pass invalid or negative values.
Consequently, DMTCP will assert and exit if it tries to get
the real ids for such invalid values.

This patch adds a case for these two commands in the semctl
wrapper to avoid the above mentioned problem.