Fix ipc close
    This removes the test of dmtcp_is_running_state().  Earlier, it would
refuse to call process_fd_event() at checkpoint time in a plugin, in
case it was opened within a plugin, and the fd was not entered into the
connection list.  But the fnc process_fd_event() will do the check later,
and will skip this if fd is not in the connection list.
    This caused a bug because the application might open a file, and
a plugin might have to close it (at checkpoint time) and re-open it
(at resume/restart).  This happened with PMI_Init() and PMI_Finalize().
We finalize before checkpointing and init again (perhaps to a different
fd) at resume/restart.  This happens in the batch-queue plugin.