threadlist.cpp: Fix compilation error for prctl
The issue arose due to a recent commit: 0523b46, in which
procname related code (from src/procname.cpp) was moved into the
stopthisthread signal handler (in src/threadlist.cpp), but the
relevant header file: <sys/prctl.h> that is required for prctl()
was never moved out of the conditional in threadlist.cpp. This
results in a compilation failure on RHEL-5.8.