This commit adds a new dmtcp API dmtcp_checkpoint_ckptDir which takes checkpoint directory name as argument and stores all checkpoint files for all processes at common place. This is necessary for DMTCP Veloce Integration as while checkpointing Veloce Software also takes HDL checkpoint which is spread across multiple sub-directories. So in order to consolidate complete database, it is necessary that all possible database related files get stored at one place. This is already present in Veloce Specific checkpointing, however missing in DMTCP application initiated checkpointing.
There is an API to set the checkpoint directory but this API
only sets checkpoint directory for the calling process. Its too
cumbersome to make all process call same API everytime when
new checkpoint activity is initiated. So we need a support which
globally changes the checkpoint directory for all processes.