fileconnection: Fix checkpointing of unlinked files opened with O_WRONLY
This patch fixes an issue where DMTCP would incorrectly restore
files that were unlinked after opening with O_WRONLY.

The policy for checkpoint-restore of such files was first defined
in the commit 5c3dd0a as follows:

 a) DMTCP should not checkpoint such files, regardless of the
    --checkpoint-open-files flags; and
 b) On restart, DMTCP should open such files with (O_WRONLY|O_CREAT)
    and unlink them.

However, the implementaiton never achieved these goals. We did end up
doing the first part correctly, but never really got the second path right.

With the above commit, files would be opened with (O_WRONLY|O_CREAT)
on restart but were never unlinked because the commit overlooked an
if-clause in the `FileConnection::resume()` method. The if-clause would
call unlink() only if a file had the `_ckpted_file` flag set to true.

Later, the commit: 8514737 worsened the situation. It introduced a bug
in the `FileConnection::refill()` method that would prevent us from even
opening write-only, unlinked files on restart.