pid-plugin: Fix incorrect initialization of pid map
This patch addresses an issue where an uninitialized shared PID map at
launch time can preclude two sibling processes from discovering each
other's real pid.

The issue can be observed when a parent that is not running under DMTCP
forks and execs child processes under DMTCP. If one of the child processes
(which is running under DMTCP) tries to look up the real pid of one of
its siblings using the sibling's virtual pid, for example, to send a
signal using the kill system call, it'll see an invalid value.  This is
because the shared PID map is never updated by the forked processes. This
issue does not affect the case when the root process itself is started
under DMTCP because the shared PID map is updated in the fork() wrapper
in `pid_miscwrappers.cpp`, which gets called when the root process tries
to fork and exec.