socketwrappers: Fix masking of real errno
This patch is an attempt to fix the issue of masking of the _real_
errno by DMTCP's socket wrappers.

The problem is that the code in a wrapper between the call to the real
function and the return statement, for example, JTRACE's, etc., can have
a side-effect of modifying the errno. As a result, the wrapper can end
up returning an incorrect errno to the user. The solution is to save
the errno once a real call returns and restore it just before returning.

Although this patch only addresses the issue for socketwrappers, the
problem is more widespread; I'm putting this out for early feedback.