ckptserializer: Fix passing a NULL environment to the gzip process
The existing implementation of the `open_ckpt_to_write()` passes a
NULL environment to the gzip child process. This can be an issue, as
reported by a user on Github (Github ID: MrPrezident), if gzip (the
real binary or a wrapper script around the real binary) relies on some
environment variables.

The fix here is to let the gzip process inherit the environment of the
parent process by using `execvp(gzip, gzip_args)` instead of `execve(gzip,
gzip_args, NULL)`. Note that the code in the function unsets LD_PRELOAD
for the child process already.