monitorSockets: Re-implement using poll()
This patch fixes an issue with the `monitorsockets()` function that
would result in a buffer overflow at checkpoint time for applications
that would open many file descriptors (> 1024). The issue arises due to
the use of `select()` call.  The root cause for the buffer overflows
is the implementation of FD_{SET|CLR|ISSET}macros in glibc. According
to the man page, it's the application's responsibility to ensure that
these macros are not called for a file-descriptor that is larger than
FD_SETSIZE, which is typically 1024 in most implementations.