pid-plugin: Fix restart callback to use _real_* functions
This appears to be an old bug. Usage of close()/dup2() can lead
to unexpected consequences where a fd associated with a different
plugin (socket, etc.) may be closed affecting their behavior on
restart.

I noticed this issue when a raw socket associated with the IPC plugin
was being closed unintentionally as a result of calling close() in the
pid plugin's restart callback. The temporary PIDMAP fd opened in the
restart callback happened to be the same as the socket fd that the raw
socket was associated with prior to checkpointing.

with '#' will be ignored, and an empty message aborts the commit.