ipc-plugin: Replace calls to select() with drain()
This patch is a follow-up to commit: 907003c. As mentioned before, the
select() call is only capable of handling 1024 fds. Usage of select()
with more than 1024 fds can lead to undefined behavior because of the
way it's implemented in libc. This patch replaces calls to select()
with poll() in four different places in the IPC plugin.