Uniquepid: Generate comp. id using higher resoulution timer
The current implementation for generating a unique computation
identifier relies on a low resolution timer function,
`time()`, with a min. resolution of 1 second. This can result in
conflicts in the computation id space when running independent
computations in parallel on the same host.

I first encountered this problem when launching independent
programs under DMTCP in parallel using `make -j`. I noticed
that a program would end up reading the coordinator info
from the shared data area of a different program because
of matching computation id values. The program would then
try to send messages to the wrong coordinator and result in
undefined behavior.

The fix is to use a higher resolution timer API, `clock_gettime()`,
with a min. resolution of 1 nanosecond.