dmtcp_coordinator: Fix a segfault when getaddrinfo() fails
The fourth argument to `getaddrinfo()`, `results`, is a linked-list that
is allocated when a call to `getaddrinfo()` succeeds. However, when the
call to the function fails, the linked-list remains unallocated and any
attempts to free the linked-list (using `freeaddrinfo()`) can result in
a segmentation fault. The patch ensures that the results linked-list is
freed up only after a successful allocation.