dmtcp_coordinator: Use higher resolution timer for generating coordTimeStamp
An issue was observed where multiple parallel launches/restarts of
independent computations on a single node would up using the incorrect
shared-data files. The reason was that since a shared-data file's name
is generated using the following tuple: <hostid, virt. pid, coordinator
timestamp>, multiple independent coordinators would end up generating
the same values when they were launched in parallel.

The fix is to use a nanosecond resolution timer for generating the
coordinator timestamp. Although this doesn't fix the issue entirely,
it does reduce the chances of two independently launched computations
on a node mapping the same shared-data file.

The patch also fixes a minor bug in the UniquePid::ThisProcess()
function.