Fixed a bug in signal handler restoration during restart.
The signal handlers must be restored before the restart hooks for
plugins are called because some plugins might perform tasks that could
potentially generate a signal. For example, the timer plugin may restore
a timer which will fire right away, and not having an appropriate signal
handler could kill the process.