Eliminate static C++ objects from jassert, jfilesystem, and dmtcpworker
This fixes issue #537. Apparently, on older rtld/libc/libstdc++, the
library acquires a low-level lock in its constructor. This can lead to
a deadlock when the library, in its constructor, ends up calling a DMTCP
wrapper, which can then call other helper functions that have their own
static local objects that need to acquire the same lock to be initialized.
Here's a sample stacktrace:

    __lll_lock_wait_private () from /lib64/libc.so.6
    _L_lock_16 () from /lib64/libc.so.6
    __new_exitfn () from /lib64/libc.so.6
    __cxa_atexit_internal () from /lib64/libc.so.6
    theLogFilePath () at ../jalib/jassert.cpp:171
    jassert_internal::set_log_file ()
    dmtcp::Util::initializeLogFile ()
    prepareLogAndProcessdDataFromSerialFile () at dmtcpworker.cpp:231
    dmtcp_initialize () at dmtcpworker.cpp:266
    calloc (nmemb=1, size=1040) at alloc/mallocwrappers.cpp:32
    __new_exitfn ()
    __cxa_atexit_internal ()
    __static_initialization_and_destruction_0
    __do_global_ctors_aux ()
    _init ()
    _dl_init_internal ()
    _dl_start_user ()