dmtcp_dlsym: Fix a crash when link_map doesn't contain vDSO entry
The commit: 3dff1b3 added a warning for the user that would be triggered
when the checkpoint and the restart machines had incompatible vDSO
layouts. The warning was added as a nice-to-have feature where earlier the
restart would have just crashed without a clue. The code would determine
the layout using the minor extension to the dmtcp_dlsym function.

However, the current implementation would lead to a crash if the vDSO
section was not listed as part of the link_map. We have now observed this
issue on two separate systems -- a user reported this in issue #566 and
@jiajuncao separately observed this on a CentOS 7.2 system. This patch
fixes this problem by ensuring that dmtcp_dlsym doesn't crash in the
corner cases.