svipc-plugin: Fix a bug in restoration of shm objects
If, during leader election, a process that didn't create an SHM object
ended up being elected the leader for checkpointing, the `_flags`
attribute of the SHM would not have the IPC_CREAT bit set. This would
cause the restart to fail when the same process, on restart, tried to
recreate the SHM object without the IPC_CREAT flag.