threadwrappers: Fix pid in TLS of newly-created threads
A bug was observed where a newly-created thread post restart would end
up reusing the TLS of a thread that existed pre-checkpoint. Libpthread
decides to recycle the old TLS, and as an optimization, would not
update the pid field in the thread-block (under the assumption that
the pid never changes). Consequently, calls to pthread_cancel(), etc.,
for the newly-created thread would end up calling system calls (tgkill,
etc.) with the wrong pid and fail.

The fix is to call `TLSInfo_UpdatePid()` for any new thread in the
`clone_start()` function, which is the first thing the new thread
executes.