Fix deadlock when a process is forked in the resume phase
We need to reset the `preResumeThreadCount` in the child process in the
atfork_child handler. Otherwise, if a process is forked in the resume
phase, this can lead to a deadlock, where the checkpoint thread that's
created for the child process in the fork() wrapper gets blocked in
waitForUserThreadsToFinishPreResumeCB(). This is because the child process
inherits the non-zero value of `preResumeThreadCount` from the parent.