Removed coordinator socket disconnect on exit.
The checkpoint thread (or some other user thread) might be in the middle
of sending a message to the coordinator. Closing the coordinator socket
might lead to unexpected error in the worker and/or coordinator. It's
better to either leave it as is or have the checkpoint thread close the
socket if/when it can.