Simplified exit() processing for ckpt-thread.
Since the ckpt-thread doesn't depend on any static C++ objects (i.e.,
DmtcpWorker singleton), we do not need to worry about the destruction of
the object while the ckpt-thread is using it. This eliminates the need
to have synchronization between ckpt-thread and the user-thread
processing exit. Instead, now the user-thread simply sets exitInProgress
and the ckpt-thread uses this information to do a clean exit.