Updated ckpt-thread checks for exit-in-progress flag.
The ckpt-thread should check for exit-in-progress only after it has
suspended user threads or when it can't read a message from coordinator
due to closed coordinator socket (all fds are closed as part of exit()
processing). If the user-threads, including the one that called exit(),
get suspended, the ckpt-thread simply resumes them so that they can run
to completion and the process can exit. The ckpt-thread then goes into
an eternal sleep.

Checking for the exit-in-progress flag at any other place makes little
sense. The only benefit of doing so would be that we'd potentially avoid
sending user-threads into the signal handler. This is a small benefit
compared to the increased complexity.