fix trampoline setup code for an edge case
Currently, the code set the page permission to 'rwx' where the
trampoline needs to be setup. This is required as we write few
bytes to the memory to set the trampoline. However, the code
segfaults if the number of bytes that need to be written reaches the
boundary of the current page and tries to write on a next read-only page.
This patch would fix the edge case by giving the `rwx` permission to
the next page if required.